StreamingDataset and MutableDataset have now deprecated warning messages. They still exists but their behaviour is directly inherited from the Dataset class, which makes them identical at this point
