+++ scrunch/__init__.py
@@ -1,8 +1,14 @@
 from .session import connect
 from .datasets import (
-    get_user, get_project, get_dataset, get_team, create_team)
+    get_user,
+    get_project,
+    get_dataset,
+    get_team,
+    create_team,
+    create_dataset
+)
 from .streaming_dataset import get_streaming_dataset
-from .mutable_dataset import get_mutable_dataset, create_dataset
+from .mutable_dataset import get_mutable_dataset
 from .version import __version__
